Desacoplar formulario de componente principal Quitar dependencias y mover a nuevo componente especifico para el formulario Componente FormItemView.jsx
